# docs/swagger.yaml
basePath: /api/v1
definitions:
  entities.Allergy:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        example: Пенициллин
        type: string
      patient:
        items:
          $ref: '#/definitions/entities.Patient'
        type: array
      updated_at:
        type: string
    type: object
  entities.ContactInfo:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      phone:
        type: string
      updated_at:
        type: string
    type: object
  entities.Doctor:
    properties:
      created_at:
        type: string
      emergency_calls:
        items:
          $ref: '#/definitions/entities.EmergencyCall'
        type: array
      id:
        type: integer
      login:
        example: doctor_ivanov
        type: string
      receptions:
        items:
          $ref: '#/definitions/entities.ReceptionHospital'
        type: array
      specialization:
        example: Терапевт
        type: string
      updated_at:
        type: string
    type: object
  entities.EmergencyCall:
    properties:
      address:
        example: Москва, ул. Ленина, д. 15
        type: string
      created_at:
        type: string
      doctor_id:
        example: 1
        type: integer
      id:
        type: integer
      phone:
        example: "+79991234567"
        type: string
      priority:
        example: true
        type: boolean
      receptions:
        items:
          $ref: '#/definitions/entities.ReceptionSMP'
        type: array
      status:
        allOf:
        - $ref: '#/definitions/entities.EmergencyStatus'
        example: in_progress
      updated_at:
        type: string
    type: object
  entities.EmergencyStatus:
    enum:
    - scheduled
    - accepted
    - on_place
    - completed
    - cancelled
    - no_show
    type: string
    x-enum-varnames:
    - EmergencyStatusScheduled
    - EmergencyStatusAccepted
    - EmergencyStatusOnPlace
    - EmergencyStatusCompleted
    - EmergencyStatusCancelled
    - EmergencyStatusNoShow
  entities.MedService:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        example: EKG
        type: string
      price:
        example: 100
        type: integer
      receptionSMP:
        items:
          $ref: '#/definitions/entities.ReceptionSMP'
        type: array
      updated_at:
        type: string
    type: object
  entities.Patient:
    properties:
      allergy:
        items:
          $ref: '#/definitions/entities.Allergy'
        type: array
      birth_date:
        example: "1980-05-15T00:00:00Z"
        type: string
      contact_info:
        $ref: '#/definitions/entities.ContactInfo'
      created_at:
        type: string
      emergency_receptions:
        items:
          $ref: '#/definitions/entities.ReceptionSMP'
        type: array
      full_name:
        example: Смирнов Алексей Петрович
        type: string
      id:
        type: integer
      is_male:
        example: true
        type: boolean
      personal_info:
        $ref: '#/definitions/entities.PersonalInfo'
      receptions:
        items:
          $ref: '#/definitions/entities.ReceptionHospital'
        type: array
      updated_at:
        type: string
    type: object
  entities.PersonalInfo:
    properties:
      created_at:
        type: string
      id:
        type: integer
      oms:
        example: "1234567890123456"
        type: string
      passport_series:
        example: 4510 123456
        type: string
      patient_id:
        example: 1
        type: integer
      snils:
        example: 123-456-789 00
        type: string
      updated_at:
        type: string
    type: object
  entities.ReceptionHospital:
    properties:
      address:
        example: Москва, ул. Ленина, д. 15
        type: string
      created_at:
        type: string
      date:
        example: "2023-10-15T14:30:00Z"
        type: string
      diagnosis:
        example: ОРВИ
        type: string
      doctor:
        $ref: '#/definitions/entities.Doctor'
      doctor_id:
        example: 1
        type: integer
      id:
        type: integer
      patient:
        $ref: '#/definitions/entities.Patient'
      patient_id:
        example: 1
        type: integer
      recommendations:
        example: Постельный режим
        type: string
      status:
        allOf:
        - $ref: '#/definitions/entities.ReceptionStatus'
        example: scheduled
      updated_at:
        type: string
    type: object
  entities.ReceptionSMP:
    properties:
      created_at:
        type: string
      diagnosis:
        example: ОРВИ
        type: string
      doctor_id:
        example: 1
        type: integer
      emergency_call_id:
        example: 1
        type: integer
      id:
        type: integer
      med_services:
        items:
          $ref: '#/definitions/entities.MedService'
        type: array
      patient_id:
        example: 1
        type: integer
      recommendations:
        example: Постельный режим
        type: string
      updated_at:
        type: string
    type: object
  entities.ReceptionStatus:
    enum:
    - scheduled
    - completed
    - cancelled
    - no_show
    type: string
    x-enum-comments:
      StatusCancelled: '"Отменен"'
      StatusCompleted: '"Завершен"'
      StatusNoShow: '"Не явился"'
      StatusScheduled: '"Запланирован"'
    x-enum-varnames:
    - StatusScheduled
    - StatusCompleted
    - StatusCancelled
    - StatusNoShow
  handlers.ResultError:
    properties:
      response:
        properties:
          code:
            description: '[RULE]: must be one of codes from table (Check DEV.PAGE)'
            example: 400
            type: integer
          message:
            example: Bad request
            type: string
        type: object
      status:
        description: error
        example: error
        type: string
    type: object
  handlers.ResultResponse:
    properties:
      response:
        properties:
          data:
            description: '[AVALIABLE]: object, array of objects, empty'
          message:
            example: Success operation
            type: string
          type:
            description: '[AVALIABLE]: object, array, empty'
            example: object
            type: string
        type: object
      status:
        description: ok
        example: ok
        type: string
    type: object
  models.AddAllergyRequest:
    description: Запрос для добавления аллергии пациенту
    properties:
      allergy_id:
        description: ID аллергии
        example: 5
        type: integer
      description:
        description: Описание аллергии
        example: Аллергия на пыльцу
        type: string
      patient_id:
        description: ID пациента
        example: 1
        type: integer
    type: object
  models.AllergyResponse:
    description: Ответ с названием аллергии
    properties:
      name:
        description: Только название аллергии
        example: Пыльца
        type: string
    type: object
  models.ContactInfoResponse:
    description: Контактная информация пациента
    properties:
      address:
        description: Физический адрес
        example: Москва, ул. Пушкина, д. 10
        type: string
      email:
        description: Электронная почта
        example: patient@example.com
        type: string
      phone:
        description: Номер телефона
        example: "+79991234567"
        type: string
    type: object
  models.CreateDoctorRequest:
    description: Используется для регистрации нового врача в системе
    properties:
      email:
        description: Email (обязательное)
        example: doctor@example.com
        type: string
      full_name:
        description: ФИО врача (обязательное)
        example: Иванов Иван Иванович
        type: string
      login:
        description: Логин (обязательное)
        example: "+79123456789"
        type: string
      password:
        description: Пароль (обязательное)
        example: qwerty123
        type: string
      specialization:
        description: Специализация (обязательное)
        example: Терапевт
        type: string
    required:
    - email
    - full_name
    - login
    - password
    - specialization
    type: object
...
  models.MedCardResponse:
    description: Содержит всю медицинскую информацию о пациенте
    properties:
      allergy:
        description: Список аллергий
        items:
          $ref: '#/definitions/models.AllergyResponse'
        type: array
      contact_info:
        allOf:
        - $ref: '#/definitions/models.ContactInfoResponse'
        description: Контактные данные
      patient:
        allOf:
        - $ref: '#/definitions/models.ShortPatientResponse'
        description: Основные данные пациента
      personal_info:
        allOf:
        - $ref: '#/definitions/models.PersonalInfoResponse'
        description: Персональная информация
    type: object
  models.PersonalInfoResponse:
    description: Документы и идентификационные данные пациента
    properties:
      oms:
        description: Номер полиса ОМС
        example: "1234567890123456"
        type: string
      passport_series:
        description: Серия и номер паспорта
        example: 4510 123456
        type: string
      snils:
        description: Номер СНИЛС
        example: 123-456-789 00
        type: string
    type: object
...
  models.ShortPatientResponse:
    description: Сокращенные данные пациента
    properties:
      birth_date:
        description: Дата рождения
        example: "1980-05-15T00:00:00Z"
        type: string
      full_name:
        description: ФИО пациента
        example: Смирнов Алексей Петрович
        type: string
      id:
        description: ID пациента
        type: integer
      is_male:
        description: Пол (true - мужской)
        example: true
        type: boolean
    type: object
  models.UpdateDoctorRequest:
    description: Используется для изменения информации о враче
    properties:
      email:
        description: Новый email
        example: doctor@example.com
        type: string
      full_name:
        description: Новое ФИО
        example: Иванов Иван Иванович
        type: string
      id:
        description: ID врача для обновления
        example: 1
        type: integer
      login:
        description: Новый логин
        example: "+79123456789"
        type: string
      password:
        description: Новый пароль
        example: newpassword123
        type: string
      specialization:
        description: Новая специализация
        example: Хирург
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API для работы с приёмами пациентов
  title: ClinicHub API
  version: 1.0.0
paths:
  /api/v1/patient/allergy:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные аллергии
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.AddAllergyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Добавленная аллергия
          schema:
            $ref: '#/definitions/models.AllergyResponse'
      summary: Добавить аллергию пациенту
      tags:
      - allergy
...
  /doctor:
    post:
      consumes:
      - application/json
      description: Создает нового врача с указанными данными
      parameters:
      - description: Данные врача
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/models.CreateDoctorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Созданный врач
          schema:
            $ref: '#/definitions/entities.Doctor'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "422":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/handlers.ResultError'
      summary: Создать нового врача
      tags:
      - Doctor
    put:
      consumes:
      - application/json
      description: Обновляет информацию о враче
      parameters:
      - description: Данные для обновления
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/models.UpdateDoctorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленный врач
          schema:
            $ref: '#/definitions/entities.Doctor'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "404":
          description: Врач не найден
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "422":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/handlers.ResultError'
      summary: Обновить данные врача
      tags:
      - Doctor
  /doctor/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет врача по ID
      parameters:
      - description: ID врача
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление
          schema:
            $ref: '#/definitions/handlers.ResultResponse'
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "404":
          description: Врач не найден
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/handlers.ResultError'
      summary: Удалить врача
      tags:
      - Doctor
    get:
      consumes:
      - application/json
      description: Возвращает информацию о враче по ID
      parameters:
      - description: ID врача
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Информация о враче
          schema:
            $ref: '#/definitions/entities.Doctor'
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "404":
          description: Врач не найден
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/handlers.ResultError'
      summary: Получить врача по ID
      tags:
      - Doctor
  /emergency/{doctor_id}/receptions:
    get:
      consumes:
      - application/json
      description: Возвращает список экстренных приёмов, назначенных врачу на указанную
        дату, с пагинацией
      parameters:
      - description: ID врача
        in: path
        name: doctor_id
        required: true
        type: integer
      - description: Дата в формате YYYY-MM-DD
        in: query
        name: date
        required: true
        type: string
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список приёмов
          schema:
            items:
              $ref: '#/definitions/entities.EmergencyCall'
            type: array
        "400":
          description: Некорректный запрос или параметры
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/handlers.ResultError'
      summary: Получить экстренные приёмы врача по дате
      tags:
      - EmergencyReception
  /medcard:
    put:
      consumes:
      - application/json
      description: Обновляет информацию о враче
      parameters:
      - description: Данные для обновления
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/models.UpdateDoctorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленный врач
          schema:
            $ref: '#/definitions/entities.Doctor'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "404":
          description: Врач не найден
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "422":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/handlers.ResultError'
      summary: Обновить данные врача
      tags:
      - Doctor
  /medcard/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает медкарту по ID Пациента
      parameters:
      - description: ID пациента
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Медкарта
          schema:
            $ref: '#/definitions/models.MedCardResponse'
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "404":
          description: Медкарта не найдена
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/handlers.ResultError'
      summary: Получить медкарту по ID Пациента
      tags:
      - MedCard
...
  /patient/{patient_id}/allergies:
    get:
      consumes:
      - application/json
      description: Возвращает список аллергий пациента
      parameters:
      - description: ID пациента
        in: path
        name: patient_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список аллергий
          schema:
            items:
              $ref: '#/definitions/models.AllergyResponse'
            type: array
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "404":
          description: Пациент не найден
          schema:
            $ref: '#/definitions/handlers.ResultError'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/handlers.ResultError'
      summary: Получить аллергии пациента
      tags:
      - Allergy
...
swagger: "2.0"